dedup to_string()
