+++ lib/python/util/__init__.py
@@ -18,3 +18,15 @@ def b64(s):
 def b64sha1sum(s):
     "Returns the base64 encoded version of the sha1sum of s"
     return b64(sha1string(s))
+
+
+def to_bytes(obj):
+    if six.PY3 and isinstance(obj, six.string_types):
+        obj = obj.encode('utf-8')
+    return obj
+
+
+def to_string(obj):
+    if six.PY3 and isinstance(obj, six.binary_type):
+        obj = obj.decode('utf-8')
+    return obj
